requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg[binary]==3.2.13
python-decouple==3.8
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.6.0
